+++ NEWS.md
@@ -2,6 +2,8 @@
 
 # development version
 
+- Add Downes 2023 vancomycin ([doi:10.3390/pharmaceutics15051336](https://doi.org/10.3390/pharmaceutics15051336)) -- critically ill children aged 1-17 years in the PICU; three models (full, eGFR-Hoek, eGFR-Schwartz).
+
 - Add Ousey 2026 plozasiran ([doi:10.1002/jcph.70190](https://doi.org/10.1002/jcph.70190)) -- adults with familial chylomicronemia syndrome.
 
 - Add Ravva 2010 varenicline exposure-response models ([doi:10.1038/clpt.2009.282](https://doi.org/10.1038/clpt.2009.282)) -- adult cigarette smokers in smoking-cessation trials.